requirements.txt
asgiref==3.4.1
astroid==2.4.2
beautifulsoup4==4.10.0
colorama==0.4.3
configparser==5.0.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django_admin_list_filter_dropdown==1.0.3
django-debug-toolbar==3.0
django-heroku==0.3.1
django-admin-rangefilter==0.8.1
django-import-export==2.3.0
django-secure==1.0.2
django-sslserver==0.22
et-xmlfile==1.0.1
gunicorn==20.0.4
isort==5.5.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.2
lxml==4.6.3
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.2
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.7.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.10.0.2
whitenoise==5.2.0
win-unicode-console==0.5
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
